update project folder structure

+++ public/admin/adminPage.php
@@ -1,20 +1,22 @@
 <?php
 
 session_start();
-require 'db.inc.php';
+require_once '../config/db.inc.php';
 
 if (!isset($_SESSION['loggedInUser'])) {
-    header('Location: login.php');
+    header('Location: ../controllers/login.php');
 }
 
 try {
-    $removeQuery = "DELETE FROM games WHERE id = :id";
-    $stmt = $conn->prepare($removeQuery);
-    $stmt->execute(
-        [
-            'id' => $_GET['id']
-        ]
-    );
+    if (isset($_GET['id'])) {
+        $removeQuery = "DELETE FROM games WHERE id = :id";
+        $stmt = $conn->prepare($removeQuery);
+        $stmt->execute(
+            [
+                'id' => $_GET['id']
+            ]
+        );
+    }
 } catch (PDOException $err) {
     $error = $err->getMessage();
 }
@@ -27,7 +29,8 @@ try {
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Admin Page</title>
-    <link rel="stylesheet" href="style.css">
+    <link rel="stylesheet" href="../css/style.css">
+    <link rel="icon" type="image/x-icon" href="../images/noBgImg.png">
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.7.2/css/all.min.css"
         integrity="sha512-Evv84Mr4kqVGRNSgIGL/F/aIDqQb7xQ2vcrdIwxfjThSH8CSR7PBEakCr51Ck+w+/U6swU2Im1vVX0SVk9ABhg=="
         crossorigin="anonymous" referrerpolicy="no-referrer">
@@ -35,20 +38,20 @@ try {
 
 <body>
     <nav class="navContainer">
-        <a href="index.php"><img src="logo.png" alt="GameNet Logo" class="logoImg"></a>
+        <a href="../index.php"><img src="../images/logo.png" alt="GameNet Logo" class="logoImg"></a>
         <ul class="navRight">
-            <li class="navList"><a href="index.php"><i class="fa-solid fa-house" style='font-size:26px'></i></a></li>
-            <li class="navList"><a href="productPage.php"><i class="fa-solid fa-shop" style='font-size:26px'></i></a></li>
-            <li class="navList"><a href="favorites.php"><i class="fa-solid fa-heart" style='font-size:26px'></i></a></li>
-            <li class="navList"><a href="cart.php"><i class="fa-solid fa-cart-shopping" style='font-size:26px'></i></a></li>
+            <li class="navList"><a href="../index.php"><i class="fa-solid fa-house" style='font-size:26px'></i></a></li>
+            <li class="navList"><a href="../productPage.php"><i class="fa-solid fa-shop" style='font-size:26px'></i></a></li>
+            <li class="navList"><a href="../user/favorites.php"><i class="fa-solid fa-heart" style='font-size:26px'></i></a></li>
+            <li class="navList"><a href="../controllers/cart.php"><i class="fa-solid fa-cart-shopping" style='font-size:26px'></i></a></li>
             <?php if (isset($_SESSION['loggedInUser'])) {
-                $navLink = 'profile.php';
+                $navLink = '../user/profile.php';
             } else {
-                $navLink = 'login.php';
+                $navLink = '../controllers/login.php';
             } ?>
             <li class="navList"><a href="<?php echo $navLink ?>"><i class='fa-solid fa-user' style='font-size:26px'></i></a></li>
             <?php if (isset($_SESSION['loggedInUser'])) {
-                ?><li class="navList"><a href="logout.php"><i class="fa-solid fa-right-from-bracket" style='font-size:26px'></i></a></li>
+            ?><li class="navList"><a href="../controllers/logout.php"><i class="fa-solid fa-right-from-bracket" style='font-size:26px'></i></a></li>
             <?php } ?>
         </ul>
     </nav>
@@ -88,11 +91,11 @@ try {
                                 <td><a href='edit.php?id=<?php echo $game->id ?>' class="hover">Wijzig product</a></td>
                             </tr>
                             <tr>
-                                <td><a href=' ?id=<?php echo $game->id ?>' class='error hover'>Verwijder</a></td>
+                                <td><a href='?id=<?php echo $game->id ?>' class='error hover'>Verwijder</a></td>
                             </tr>
                         </table>
                     </td>
-                    <?php
+            <?php
                     $counter++;
 
                     if ($counter % 4 == 0) {
@@ -117,7 +120,7 @@ try {
 
     <footer class="footer-container">
         <p>Copyright &copy; GameNet 2025</p>
-        <p><a href="about.php">About</a></p>
+        <p><a href="../about.php">About</a></p>
     </footer>
 </body>
